[BAEL-16045] - Fixed article code for https://www.baeldung.com/get-user-in-spring-security
