+++ spring-security-rest-custom/src/main/java/org/baeldung/web/controller/SecurityController.java
@@ -2,26 +2,23 @@ package org.baeldung.web.controller;
 
 import java.security.Principal;
 
-import javax.servlet.http.HttpServletRequest;
-
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
-public class SecurityController4 {
+public class SecurityController {
 
-    public SecurityController4() {
+    public SecurityController() {
         super();
     }
 
     // API
 
-    @RequestMapping(value = "/username4", method = RequestMethod.GET)
+    @RequestMapping(value = "/username2", method = RequestMethod.GET)
     @ResponseBody
-    public String currentUserNameSimple(final HttpServletRequest request) {
-        final Principal principal = request.getUserPrincipal();
+    public String currentUserNameSimple(final Principal principal) {
         return principal.getName();
     }
